==增加一些容错判断

+++ Client/Assets/Game/YouYouFramework/Editor/YouYouMenuExt.cs
@@ -144,18 +144,22 @@ public class YouYouMenuExt
     [MenuItem("YouYouTools/生成并拷贝热更新程序集到Download文件夹")]
     public static void CompiltHotifxDll()
     {
-        Debug.Log("生成并拷贝热更新程序集到Download文件夹");
-        HybridCLR.Editor.Commands.CompileDllCommand.CompileDll(EditorUserBuildSettings.activeBuildTarget);
-
-        string CodeDir = "Assets/Game/Download/Hotfix/";
-
-        string ScriptAssembliesDir = Application.dataPath + "/../" + "HybridCLRData/HotUpdateDlls/" + EditorUserBuildSettings.activeBuildTarget.ToString() + "/Assembly-CSharp.dll";
-        File.Copy(ScriptAssembliesDir, Path.Combine(CodeDir, "Assembly-CSharp.dll.bytes"), true);
-
         string aotMetaAssemblyDir = Application.dataPath + "/../" + "HybridCLRData/AssembliesPostIl2CppStrip/" + EditorUserBuildSettings.activeBuildTarget + "/";
-        foreach (var aotDllName in HotfixCtrl.aotMetaAssemblyFiles)
+        if (Directory.Exists(aotMetaAssemblyDir))
         {
-            File.Copy(aotMetaAssemblyDir + aotDllName + ".dll", Path.Combine(CodeDir, aotDllName + ".dll.bytes"), true);
+            Debug.Log("生成并拷贝热更新程序集到Download文件夹");
+            string CodeDir = "Assets/Game/Download/Hotfix/";
+
+            //生成热更程序集
+            HybridCLR.Editor.Commands.CompileDllCommand.CompileDll(EditorUserBuildSettings.activeBuildTarget);
+            string ScriptAssembliesDir = Application.dataPath + "/../" + "HybridCLRData/HotUpdateDlls/" + EditorUserBuildSettings.activeBuildTarget.ToString() + "/Assembly-CSharp.dll";
+            File.Copy(ScriptAssembliesDir, Path.Combine(CodeDir, "Assembly-CSharp.dll.bytes"), true);
+
+            //生成AOT补充程序集
+            foreach (var aotDllName in HotfixCtrl.aotMetaAssemblyFiles)
+            {
+                File.Copy(aotMetaAssemblyDir + aotDllName + ".dll", Path.Combine(CodeDir, aotDllName + ".dll.bytes"), true);
+            }
         }
     }
     #endregion
